feat: add FormattedAddressBlock component to display user address
